refactor: separate concerns across cli / application / infra layers
- cli/prompt.rs: dispatch only (run_refresh/run_direct/run_cached moved out)
- application/prompt.rs: use case orchestration (run_cached, run_refresh,
  run_direct, fetch_silent + TrackingPresenter)
- infra/refresh_lock.rs: lock management + maybe_spawn_refresh (extracted
  from cli/prompt.rs; create_with_pid helper removes duplicate code)
- infra/cache.rs: CacheStore read/write only (lock logic removed)
