+++ src/infra.rs
@@ -1,4 +1,5 @@
 pub mod cache;
 pub mod gh;
 pub mod logger;
+pub mod refresh_lock;
 pub mod repo_id;
