+++ src/infra/cache.rs
@@ -1,5 +1,4 @@
 use std::fs;
-use std::io::Write as _;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use serde::{Deserialize, Serialize};
@@ -84,93 +83,6 @@ pub fn write(repo_id: &str, output: &str) {
     };
 
     let _ = fs::write(state_path, content);
-}
-
-/// リフレッシュロックを取得する。成功時は `true`、既に起動中なら `false` を返す。
-///
-/// ロックファイルが存在する場合は PID で生存確認（`kill -0`）を行い、
-/// プロセスが死んでいれば stale とみなして除去し再取得する。
-///
-/// ロック取得直後に自プロセスの PID を書き込むため、ロックファイルが空になる瞬間は存在しない。
-/// 空ファイル＝クラッシュ等で書き込まれなかった異常状態として「死んでいる」と判定する。
-pub fn try_acquire_refresh_lock(repo_id: &str) -> bool {
-    let Some(path) = lock_path(repo_id) else {
-        return false;
-    };
-    if let Some(parent) = path.parent() {
-        let _ = fs::create_dir_all(parent);
-    }
-
-    if let Ok(mut f) = fs::OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(&path)
-    {
-        // 取得直後に自 PID を書き込む（空ファイル期間をなくす）
-        let _ = f.write_all(std::process::id().to_string().as_bytes());
-        return true;
-    }
-
-    // ロックファイルが既存: 生存確認して死んでいれば再取得
-    if is_lock_alive(&path) {
-        return false;
-    }
-    let _ = fs::remove_file(&path);
-    if let Ok(mut f) = fs::OpenOptions::new()
-        .write(true)
-        .create_new(true)
-        .open(&path)
-    {
-        let _ = f.write_all(std::process::id().to_string().as_bytes());
-        true
-    } else {
-        false
-    }
-}
-
-/// spawn 後に子プロセスの PID をロックファイルへ上書きする。
-pub fn update_lock_pid(repo_id: &str, pid: u32) {
-    if let Some(path) = lock_path(repo_id) {
-        let _ = fs::write(path, pid.to_string());
-    }
-}
-
-/// リフレッシュロックを解放する。
-pub fn release_refresh_lock(repo_id: &str) {
-    if let Some(path) = lock_path(repo_id) {
-        let _ = fs::remove_file(path);
-    }
-}
-
-/// ロックファイルが示すプロセスが生存しているかを確認する。
-///
-/// - PID あり → `kill -0 <pid>` でプロセス生存確認
-/// - PID なし（空ファイル）→ 異常状態として dead 扱い
-fn is_lock_alive(path: &std::path::Path) -> bool {
-    let content = fs::read_to_string(path).unwrap_or_default();
-    let trimmed = content.trim();
-
-    if trimmed.is_empty() {
-        return false;
-    }
-
-    trimmed.parse::<u32>().is_ok_and(|pid| {
-        std::process::Command::new("kill")
-            .args(["-0", &pid.to_string()])
-            .stderr(std::process::Stdio::null())
-            .status()
-            .is_ok_and(|s| s.success())
-    })
-}
-
-fn lock_path(repo_id: &str) -> Option<std::path::PathBuf> {
-    let home = std::env::var_os("HOME")?;
-    Some(
-        std::path::Path::new(&home)
-            .join(".cache")
-            .join(CACHE_DIR_NAME)
-            .join(format!("{repo_id}.lock")),
-    )
 }
 
 fn cache_path(repo_id: &str) -> Option<std::path::PathBuf> {
